Step 2E: Pokemon detail UI (ViewModel + Compose screen + favorite toggle)
Domain refactor (architecture cleanup):
- PokemonDetail.height/weight: Int -> Double, now stored as SI units
  (meters/kg) instead of API raw decimeters/hectograms. Mapper handles
  the /10 conversion. Keeps API conventions out of the domain layer.

ViewModel (PokemonDetailViewModel):
- @HiltViewModel injecting SavedStateHandle to extract the navigation
  argument id (Int or String fallback for robustness).
- Uses GetPokemonDetailUseCase to load the detail (suspend -> Result).
- Exposes uiState: StateFlow<UiState<PokemonDetail>>.
- Exposes favoriteState: StateFlow<Boolean> via IsFavoriteUseCase().stateIn,
  reactively driven by Room — heart icon flips automatically on toggle.
- onToggleFavorite() reads current favorite state, delegates to
  ToggleFavoriteUseCase which add/remove based on it.

Screen (PokemonDetailScreen.kt):
- PokemonDetailRoute (stateful) + PokemonDetailScreen (stateless) pattern.
- Scaffold + TopAppBar with back arrow + FAB for favorite toggle.
- FAB color and icon switch (Favorite/FavoriteBorder) based on isFavorite.
- when(uiState) exhaustive for Loading/Error/Success.
- DetailContent: vertical scroll with #id, large sprite, type chips,
  height/weight info row, stats section with 6 StatBar.
- Error state has a retry button (consistent with list screen).

Components/StatBar.kt:
- Reusable label + value + LinearProgressIndicator capped at 255 (Pokemon
  max base stat).
- @Preview function for split view in Android Studio.

+++ app/src/main/java/com/example/pokedex/domain/model/PokemonDetail.kt
@@ -5,22 +5,21 @@ package com.example.pokedex.domain.model
  * Compose un [Pokemon] de base + les informations spécifiques au détail.
  *
  * @property pokemon   Infos de base (réutilisé tel quel depuis la liste)
- * @property height    Taille en décimètres (unité PokéAPI). 7 = 0.7m.
- * @property weight    Poids en hectogrammes (unité PokéAPI). 60 = 6.0kg.
+ * @property height    Taille en **mètres** (déjà converti depuis les décimètres de l'API)
+ * @property weight    Poids en **kilogrammes** (déjà converti depuis les hectogrammes de l'API)
  * @property abilities Capacités ("static", "lightning-rod", ...)
  * @property stats     Statistiques de base
  */
 data class PokemonDetail(
     val pokemon: Pokemon,
-    val height: Int,
-    val weight: Int,
+    val height: Double,
+    val weight: Double,
     val abilities: List<String>,
     val stats: PokemonStats
 )
 
 /**
- * Statistiques de base d'un Pokémon.
- * Champs nommés explicitement (plutôt que Map<String, Int>) pour la sécurité type.
+ * Statistiques de base d'un Pokémon (valeur max théorique : 255).
  */
 data class PokemonStats(
     val hp: Int,
